+++ lazy_tags/templatetags/lazy_tags.py
@@ -1,5 +1,4 @@
 import json
-import uuid
 
 from django import template
 from django.template.loader import render_to_string
@@ -7,14 +6,14 @@ from django.conf import settings
 from django.core.cache import cache
 from django.core.urlresolvers import reverse
 
-from ..utils import set_lazy_tags_cache
+from ..utils import get_tag_id, set_lazy_tags_cache
 
 
 register = template.Library()
 
 
-@register.simple_tag(takes_context=True)
-def lazy_tag(context, tag, *args, **kwargs):
+@register.simple_tag
+def lazy_tag(tag, *args, **kwargs):
     """
     Lazily loads a template tag after the page has loaded. Requires jQuery
     (for now).
@@ -31,15 +30,8 @@ def lazy_tag(context, tag, *args, **kwargs):
         *args: arguments to be passed to the template tag.
         **kwargs:  keyword arguments to be passed to the template tag.
     """
-    tag_id = str(uuid.uuid4())
+    tag_id = get_tag_id()
     set_lazy_tags_cache(tag_id, tag, args, kwargs)
-
-    # Gross hack to pollute the parent context so lazy_tag_ids can be accessed
-    # from the lazy_tags_javascript tag
-    c = context.dicts[0]
-    if not c.get('lazy_tag_ids'):
-        c['lazy_tag_ids'] = []
-    c['lazy_tag_ids'].append(tag_id)
 
     return render_to_string('lazy_tags/lazy_tag.html', {
         'tag_id': tag_id,
@@ -47,15 +39,35 @@ def lazy_tag(context, tag, *args, **kwargs):
     })
 
 
-@register.simple_tag(takes_context=True)
-def lazy_tags_js(context):
-    """Outputs the necessary JavaScript to load tags over AJAX."""
-    lazy_tag_ids = context.get('lazy_tag_ids')
-    error_message = 'An error occurred.'
-    if hasattr(settings, 'LAZY_TAGS_ERROR_MESSAGE'):
-        error_message = getattr(settings, 'LAZY_TAGS_ERROR_MESSAGE')
+def _render_js(library):
+    error_message = getattr(settings, 'LAZY_TAGS_ERROR_MESSAGE', 'An error occurred.')
+    template = 'lazy_tags/lazy_tags_{0}.html'.format(library)
 
-    return render_to_string('lazy_tags/lazy_tags_js.html', {
-        'lazy_tag_ids': lazy_tag_ids,
+    return render_to_string(template, {
         'error_message': error_message,
     })
+
+
+@register.simple_tag
+def lazy_tags_javascript():
+    """Outputs the necessary JavaScript to load tags over AJAX."""
+    return _render_js('javascript')
+
+
+@register.simple_tag
+def lazy_tags_jquery():
+    """Outputs the necessary jQuery to load tags over AJAX."""
+    return _render_js('jquery')
+
+
+@register.simple_tag
+def lazy_tags_prototype():
+    """Outputs the necessary Prototype to load tags over AJAX."""
+    return _render_js('prototype')
+
+
+@register.simple_tag
+def lazy_tags_js():
+    """An alias to the JavaScript library specified in settings."""
+    library = getattr(settings, 'LAZY_TAGS_AJAX_JS', 'jquery')
+    return _render_js(library.lower())